Add MNASolver implementation that uses a plugin provided as .so
This adds the MNASolverPlugin Implementation of MNASolver that
uses a plugin provieded as a .so library to solve the system.
This library is loaded dynamically during runtime. The library
implementation can thus be changes without recompiling or relinking
dpsim. The library has to implement the interface specified in
MNASolvAerDynInterface.h.

This is a proof-of-concept implementation. Currently
MNASolverPlugin always loads an library called "plugin.so" which
should must be present in LD_LIBRARY_PATH. Logging is only partly
possible, and proper error handling ist still missing.
Also this only supports sparse solvers for now, and there may be
some parameters missing in the interface.

Signed-off-by: Niklas Eiling <niklas.eiling@eonerc.rwth-aachen.de>